% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecv.R
\name{generate_binary_data}
\alias{generate_binary_data}
\title{Generate binary data example}
\usage{
generate_binary_data(n = 100, p = 50, q = 3)
}
\arguments{
\item{n}{Number of observations}

\item{p}{Number of variables}

\item{q}{True number of factors}
}
\description{
Generate simulated data based on Binary factor model
}
