% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecv.R
\name{generate_continuous_data}
\alias{generate_continuous_data}
\title{Generate continuous data example}
\usage{
generate_continuous_data(n = 100, p = 50, q = 3, noise_sd = 1)
}
\arguments{
\item{n}{Number of observations}

\item{p}{Number of variables}

\item{q}{True number of factors}

\item{noise_sd}{The variance of noise}
}
\description{
Generate simulated data based on Gaussian factor model
}
