% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecv_miss.R
\name{generate_continuous_data_miss}
\alias{generate_continuous_data_miss}
\title{Generate continuous data with missing values}
\usage{
generate_continuous_data_miss(
  n = 100,
  p = 50,
  q = 3,
  noise_sd = 1,
  miss_prop = 0.05
)
}
\arguments{
\item{n}{Number of observations}

\item{p}{Number of variables}

\item{q}{True number of factors}

\item{noise_sd}{Standard deviation of noise}

\item{miss_prop}{Proportion of missing values (default 0.05)}
}
\description{
Generate simulated data based on Gaussian factor model with missing values
}
